collision handling extracted out of Monster.cpp and Player.cpp such that a new CollisionHandler.cpp class is responsible for collision logic

+++ src/Game/Entities/Monsters/Boar.cpp
@@ -46,7 +46,7 @@ void Boar::MaybeOink() {
     if (timeSinceLastOink_ >= nextOinkTime_) { // Check if the time for the next oink has passed
         Notify(this, EVENT_BOAR_OINKED);
         timeSinceLastOink_ = 0.0f; // Reset the timer
-        nextOinkTime_ = (rand() % 30) + 1; // Set the time for the next oink to a random value between 1 and 30 seconds
+        nextOinkTime_ = (float)(rand() % 30) + 1; // Set the time for the next oink to a random value between 1 and 30 seconds
     }
 }
